# backend/openapi.yaml
openapi: 3.0.3
info:
  title: 云听音频精修系统 API
  description: |
    云听音频精修系统完整 API 文档
    
    本系统提供以下功能模块：
    - 任务管理：创建任务、获取任务详情
    - 拆句管理：获取拆句列表、删除拆句
    - 合成管理：合成断句、拆句、任务，设置合成参数，查询合成状态
    - 阅读规范：创建阅读规范、获取列表、匹配字段
    - 语音配置：获取语音列表
    - 音频合并：合并音频、查询合并状态
    - 健康检查：服务健康状态
    
    所有接口统一使用 ApiResponse 包装返回数据。
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/task/create:
    post:
      tags:
        - 任务管理
      summary: 创建任务
      description: |
        创建任务时会自动进行拆句处理（根据指定的拆句符号拆句）
        拆句完成后，每条拆句记录默认会生成一条断句记录
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskCreateResponseDTO'
  /api/task/getDetail:
    get:
      tags:
        - 任务管理
      summary: 获取任务详情
      description: 返回任务详情以及任务下的拆句内容
      operationId: getTaskDetail
      parameters:
        - name: taskid
          in: query
          required: true
          description: 任务ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskDetailDTO'
  /api/originalSentence/getOriginalSentenceList:
    get:
      tags:
        - 拆句管理
      summary: 获取拆句列表
      description: 获取指定任务下的拆句列表，支持分页
      operationId: getOriginalSentenceList
      parameters:
        - name: taskid
          in: query
          required: true
          description: 任务ID
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          required: false
          description: 每页大小
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_OriginalSentenceListResponseDTO'
  /api/originalSentence/delete:
    post:
      tags:
        - 拆句管理
      summary: 删除拆句
      description: 删除指定的拆句
      operationId: deleteOriginalSentence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OriginalSentenceDeleteRequest'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/breakingSentence:
    post:
      tags:
        - 合成管理
      summary: 合成断句
      description: 合成或重新合成单个断句
      operationId: synthesizeBreakingSentence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisBreakingSentenceRequest'
      responses:
        '200':
          description: 合成请求已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/originalSentence:
    post:
      tags:
        - 合成管理
      summary: 合成拆句
      description: 合成或重新合成拆句下的所有断句
      operationId: synthesizeOriginalSentence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisOriginalSentenceRequest'
      responses:
        '200':
          description: 合成请求已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/task:
    post:
      tags:
        - 合成管理
      summary: 合成任务
      description: 合成或重新合成任务下的所有断句
      operationId: synthesizeTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisTaskRequest'
      responses:
        '200':
          description: 合成请求已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/setConfig:
    post:
      tags:
        - 合成管理
      summary: 设置拆句合成参数
      description: 设置拆句的合成参数，覆盖旧数据
      operationId: setConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SynthesisSetConfigRequest'
      responses:
        '200':
          description: 配置更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/callback:
    post:
      tags:
        - 合成管理
      summary: 华为云TTS回调接口
      description: |
        接收华为云TTS异步任务的回调通知
        将回调请求发送到RocketMQ消息队列，由消费者异步处理
      operationId: handleTtsCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TtsCallbackRequest'
      responses:
        '200':
          description: 回调消息已发送到消息队列
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/synthesis/getBreakingSentenceStatus:
    get:
      tags:
        - 合成管理
      summary: 获取断句合成状态
      description: |
        给出单句断句的合成状态和已完成合成的音频文件下载地址和时长
      operationId: getBreakingSentenceStatus
      parameters:
        - name: breakingSentenceId
          in: query
          required: true
          description: 断句ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SynthesisResultDTO'
  /api/synthesis/getOriginalSentenceStatus:
    get:
      tags:
        - 合成管理
      summary: 获取拆句合成状态
      description: |
        给出拆句下所有断句的合成进度和已完成合成的音频文件下载地址和时长
      operationId: getOriginalSentenceStatus
      parameters:
        - name: originalSentenceId
          in: query
          required: true
          description: 拆句ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_OriginalSentenceSynthesisStatusDTO'
  /api/synthesis/getTaskStatus:
    get:
      tags:
        - 合成管理
      summary: 获取任务合成状态
      description: |
        给出任务下所有断句的合成进度和已完成合成的音频文件下载地址和时长
      operationId: getTaskStatus
      parameters:
        - name: taskid
          in: query
          required: true
          description: 任务ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TaskSynthesisStatusDTO'
  /api/readingRule/create:
    post:
      tags:
        - 阅读规范
      summary: 创建阅读规范
      description: 创建新的阅读规范
      operationId: createReadingRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingRuleCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReadingRuleCreateResponseDTO'
  /api/readingRule/setGlobalSetting:
    post:
      tags:
        - 阅读规范
      summary: 开关全局阅读规范
      description: 设置全局阅读规范的开关状态
      operationId: setGlobalSetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingRuleSetGlobalSettingRequest'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/readingRule/getList:
    get:
      tags:
        - 阅读规范
      summary: 获取阅读规范列表
      description: 获取阅读规范列表（支持分页和筛选）
      operationId: getReadingRuleList
      parameters:
        - name: task_id
          in: query
          required: false
          description: 任务ID（可选）
          schema:
            type: integer
            format: int64
        - name: ruleType
          in: query
          required: false
          description: 规则类型（可选）
          schema:
            type: integer
        - name: page
          in: query
          required: true
          description: 页码（必填）
          schema:
            type: integer
        - name: pageSize
          in: query
          required: true
          description: 每页大小（必填）
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReadingRuleListPageResponseDTO'
  /api/readingRule/getMatchingFieldListFromText:
    get:
      tags:
        - 阅读规范
      summary: 获取文本中符合规则的字段列表
      description: 从文本中提取符合阅读规则的字段列表
      operationId: getMatchingFieldListFromText
      parameters:
        - name: text
          in: query
          required: false
          description: 文本内容（可选）
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MatchingFieldListResponseDTO'
...
components:
  schemas:
...
    ApiResponse_String:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，10200表示成功
          example: 10200
        message:
          type: string
          description: 响应消息
          example: success
        data:
          type: string
          description: 响应数据
        timestamp:
          type: integer
          format: int64
          description: 时间戳（毫秒）
    ApiResponse_TaskCreateResponseDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TaskCreateResponseDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_TaskDetailDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TaskDetailDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_OriginalSentenceListResponseDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/OriginalSentenceListResponseDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_SynthesisResultDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/SynthesisResultDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_OriginalSentenceSynthesisStatusDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/OriginalSentenceSynthesisStatusDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_TaskSynthesisStatusDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TaskSynthesisStatusDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_ReadingRuleCreateResponseDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ReadingRuleCreateResponseDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_ReadingRuleListPageResponseDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ReadingRuleListPageResponseDTO'
        timestamp:
          type: integer
          format: int64
    ApiResponse_MatchingFieldListResponseDTO:
      type: object
      properties:
        code:
          type: integer
          example: 10200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/MatchingFieldListResponseDTO'
        timestamp:
          type: integer
          format: int64
...
    TaskCreateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 10000
          description: 文本内容，不能为空，最多10000字
          example: 这是一段需要转换为语音的文本内容。支持最多10000字。
        delimiterList:
          type: array
          items:
            type: integer
          description: |
            自定义拆句符号类型数组，可选；为空时使用默认符号集
            1：中文句号（。）
            2：中文叹号（！）
            3：中文问号（？）
            4：中文省略号（…）
          example: [1, 2, 3]
    TaskCreateResponseDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
          description: 任务ID
          example: 1
        content:
          type: string
          description: 任务文本内容
          example: 这是一段需要转换为语音的文本内容。支持最多10000字。
        charCount:
          type: integer
          description: 字符数
          example: 28
        status:
          type: integer
          description: 任务状态
          example: 2
        audioUrl:
          type: string
          nullable: true
          description: 音频URL
          example: null
        audioDuration:
          type: integer
          nullable: true
          description: 音频时长（秒）
          example: null
        originalSentenceList:
          type: array
          description: 拆句列表
          items:
            $ref: '#/components/schemas/OriginalSentenceDTO'
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: 2024-01-01T00:00:00
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: 2024-01-01T00:00:00
    TaskDetailDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
          description: 任务ID
          example: 1
        mergeId:
          type: integer
          format: int64
          nullable: true
          description: 合并任务ID
          example: null
        content:
          type: string
          description: 任务文本内容
          example: 这是一段需要转换为语音的文本内容。
        charCount:
          type: integer
          description: 字符数
          example: 18
        status:
          type: integer
          description: 任务状态
          example: 2
        audioUrl:
          type: string
          nullable: true
          description: 音频URL
          example: null
        audioDuration:
          type: integer
          nullable: true
          description: 音频时长（秒）
          example: null
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: 2024-01-01T00:00:00
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: 2024-01-01T00:00:00
...
    OriginalSentenceListResponseDTO:
      type: object
      properties:
        list:
          type: array
          description: 拆句列表
          items:
            $ref: '#/components/schemas/OriginalSentenceListItemDTO'
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 10
        page:
          type: integer
          description: 当前页码
          example: 1
        pageSize:
          type: integer
          description: 每页大小
          example: 10
...
    OriginalSentenceDeleteRequest:
      type: object
      required:
        - originalSentenceId
      properties:
        originalSentenceId:
          type: integer
          format: int64
          description: 拆句ID
          example: 1
    SynthesisBreakingSentenceRequest:
      type: object
      required:
        - breakingSentenceId
      properties:
        breakingSentenceId:
          type: integer
          format: int64
          description: 断句ID
          example: 1
    SynthesisOriginalSentenceRequest:
      type: object
      required:
        - originalSentenceId
      properties:
        originalSentenceId:
          type: integer
          format: int64
          description: 拆句ID
          example: 1
    SynthesisTaskRequest:
      type: object
      required:
        - taskId
      properties:
        taskId:
          type: integer
          format: int64
          description: 任务ID
          example: 1
    SynthesisSetConfigRequest:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
          description: 任务ID
          example: 1
        originalSentenceId:
          type: integer
          format: int64
          description: 拆句ID
          example: 1
        breakingSentenceList:
          type: array
          description: 断句配置列表
          items:
            $ref: '#/components/schemas/BreakingSentenceConfig'
    BreakingSentenceConfig:
      type: object
      properties:
        breakingSentenceId:
          type: integer
          format: int64
          description: 断句ID
          example: 1
        sequence:
          type: integer
          description: 序号
          example: 1
        content:
          type: string
          description: 断句内容
          example: 这是一段需要转换为语音的文本内容。
        volume:
          type: integer
          description: 音量（0-100）
          example: 50
        voiceId:
          type: string
          description: 语音ID
          example: voice_001
        speed:
          type: integer
          description: 语速
          example: 0
        breakList:
          type: array
          description: 停顿配置列表
          items:
            $ref: '#/components/schemas/BreakConfig'
        phonemeList:
          type: array
          description: 多音字配置列表
          items:
            $ref: '#/components/schemas/PhonemeConfig'
        prosodyList:
          type: array
          description: 局部语速配置列表
          items:
            $ref: '#/components/schemas/ProsodyConfig'
        silenceList:
          type: array
          description: 静音配置列表
          items:
            $ref: '#/components/schemas/SilenceConfig'
        readRule:
          type: array
          description: 阅读规范配置列表
          items:
            $ref: '#/components/schemas/ReadRuleConfig'
...
    SynthesisResultDTO:
      type: object
      properties:
        audioUrl:
          type: string
          nullable: true
          description: 音频URL
          example: https://example.com/audio.mp3
        audioDuration:
          type: integer
          nullable: true
          description: 音频时长（秒）
          example: 5
        synthesisStatus:
          type: integer
          description: 合成状态（0-未合成，1-合成中，2-已合成，3-合成失败）
          example: 2
...
    TtsCallbackRequest:
      type: object
      properties:
        status:
          type: string
          description: |
            任务状态
            FINISHED - 已完成
            ERROR - 失败
            WAITING - 等待中
          example: FINISHED
        job_id:
          type: string
          description: 任务ID（华为云返回的job_id）
          example: job_123456
        audio_file_download_url:
          type: string
          nullable: true
          description: 音频文件下载URL
          example: https://example.com/audio.mp3
        subtitle_file_download_url:
          type: string
          nullable: true
          description: 字幕文件下载URL（可选）
          example: https://example.com/subtitle.srt
        audio_duration:
          type: integer
          nullable: true
          description: 音频时长（秒）
          example: 5
    ReadingRuleCreateRequest:
      type: object
      required:
        - pattern
        - ruleType
        - ruleValue
      properties:
        pattern:
          type: string
          description: 原始词（匹配模式）
          example: "\\d+"
        ruleType:
          type: integer
          minimum: 1
          maximum: 3
          description: |
            规范类型：
            1：数字英文
            2：音标调整
            3：专有词汇
          example: 1
        ruleValue:
          type: string
          description: |
            自定义读法，类型为SAY_AS时，只允许传下面的值。
            number：数字
            date：日期
            figure：数值
            phone：电话号码
            english：英文单词
            spell：逐个字母读英文
          example: "number"
...
    ReadingRuleSetGlobalSettingRequest:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
          description: 任务ID
          example: 1
        ruleId:
          type: integer
          format: int64
          description: 规则ID
          example: 1
        isOpen:
          type: boolean
          description: 是否开启（0关闭，1打开）
          example: true
        breakingSentenceId:
          type: integer
          format: int64
          nullable: true
          description: |
            断句ID（可选），如果有值则type=1（任务级），如果为空则type=2（断句级）
          example: null
...